ZMO-92: add genre to model and view
the genre is attributed to the article but the text to display is found in an external xml in the config directory. there the prose text is defined. additionally some genres a set in the xml but not displayed in the frontend yet, so there is a boolish attribute for this