add filter for image-urls
image node have an attribute base-url, from which the path to the image-server can be extracted. this filter does this in a small part, as this will be a feature for the file connector later. images are always processed as jpg and 540x304 px size (which will work in 80% of all cases)