ZMO-362: fix bug where biga height is zero
Due to hight loading speed the height of an inline gallery is now
calculated before the images are relaoded from the noscript elem.
Added an event to the rescale_one function and a listener to the
gallery which on ocasion redraws the slider.

If you'll find problems with calculated heights of responsive
images elsewhere, just set up a listener for "scaling_ready" to a
parental element of your image.