ZON-1078: Refactoring: add new default templates
To make the teaser templates comply with BEM structure, we need to refactor the default templates. The new ones are found inside this commit, please update your teaser templates to make use of them. When all teasers are refactored, the old defaults should be replaced and the new renamed and used instead. Much ado, but neccessary.