Fix infobox: Change obnoxious infobox source
Ok, using form elements for switching tabs was a good idea but not ready for the world or the requirement for a tab interface and a accordion. So we drop the double navigation and do this with a script later…

Changes the noscript fallback to show the whole infobox content instead of a working infobox.