HOTFIX: Don't break in auto areas when videos are teasered on a CP
In zeit.web it's possible to adapt a Video to IImageGroup, thus the
XMLReferenceUpdater does it, so we need to conform to the interface
enough for its needs, else it breaks.

I'm not sure whether we shouldn't rather try to make
XMLReferenceUpdater *not do this* in the first place, since the
results will be... misleading at best, but I don't see a quick way to
do that, so we leave it at that for now.