MAINT: Use dynamically-generated HTTP ports for tests
This allows us to have multiple testruns (e.g. Jenkins jobs) in parallel.

To make this work, we also need two additional changes, which makes
this commit a bit messy:

1. Use a custom "static_view" for fixture `mockserver` that replaces
%HOST% with the actual host/port, so e.g. mocked spektrum responses
can contain full URLs with the generated port.

2. Register "something else" as ISettings. (This change will be
helpful for the tests in other ways, too.)
Previously we registered an isolated dict as ISettings, which especially
had nothing to do with Pyramid's `request.registry.settings`.
But that meant that the tests had a) no _one_ way to get at the settings
and b) no easy means of manipulating Pyramid's settings.

Now we register Pyramid's settings dict as ISettings, so tests have
uniform access to it. Since we have some non-Pyramid utilities too,
which are instantiated before Pyramid and also might need to access
ISettings, this necessitated a workaround.

2a. Extract the "settings reset" between tests into a separate
fixture, so fixtures other than `application` can use it.