Serie-level42: Read level4 from banner mapping
Look, we all we keen to know what level4 will be one day. Now the mistery is solved: it has no particular meaning, as level2 and level3 alike. With this in mind one could think, the complete adcontroller info object makes no sense at all. Save this for later: this will and has to break one fine day and it's not our fault.