ZON-2925: Don't needlessly load comment thread in comment form ESI
Split view variables so that the comment-form template only touches
those it really needs.

At the moment this does not matter much, since the thread is cached by
displaying the comment area already, but with ZON-2851 we'll remove
that caching, and then we should not generate unnecessary requests to
the community.