ZON-2894: Set up ESI processing for selenium tests
Since wesgi implements ESI slightly differently than Varnish does, we
don't want to rely on it very much in tests. But the comment area uses
javascript, so we need the _whole combined_ page (i.e. the o-wrap head
and the ESI-included comments) in order to really test it.

For this to work we switch from wsgiref to waitress as the http
server, since it can handle more than one request simultaneously,
which is of course necessary for ESI processing.