BUG-593: Make multiple videos possible
The videostage plugin was programmed to listen to a range of clicktargets (although there is only one in a videostage in the moment), but would chose only the first one in DOM (rows 12 - 28). By finding the dom element after the click, relativ to the events this, multiple videos would be possible.