ZON-2284: This test is weird
why would we want to pass on rendering if we actually can render it?
Also, if you check http://friedbert.devel.zeit.de/zeit-online/render-with-header
running at zeit.web 2.98 (mid 2016) you'll see that it renders perfectly
fine. So either the test and/or the "feature" never worked or I'm missing the
point ¯\_(ツ)_/¯