OPS-495: Configure our custom error swallowing Jinja enviroment properly
Our error swallowing feature for production consists of two parts:
* Custom Environment class that swallows all exceptions instead of letting them raise
* Custom Undefined class that does not raise UndefinedError

The idea is that we configure our custom environment class, and it
pulls in the custom Undefined class automatically.

But pyramid_jinja2 has a configuration settings for the Undefined
class, and therefore explicitly passes the classname in as a keyword
parameter. Thus we have to change it explicitly, not just use a
default argument. (Previously tests only instantiated the environment
directly and thus didn't catch this issue.)