ZON-3788: Add redirect step through premium for gated articles
This is a workaround for the architectural deficiency that the
accountserver does not have an eventbus where interested parties can
subscribe and then act on events like "user was confirmed". Thus
premium needs the user to *visit* there after the confirmation, so it
can perform actions like letting C1 know that this user now has an abo.