ZON-3757: Remove overcomplicated banner listing
The inclusion of data from banner.xml was needed to inject old styled banner before we switched to the adcontroller. From that point on, a place only needed it’s tilenumber and adtype (mobile or desktop) to work. Throw away all other data, including file loading, making of a global banner_list and extra code for old school mobile banner.