BUG-736: Trying to guess if the safari bug applies
As long as Safari 10 has a bug in the font loading api that:
a) exposes different font sizes during load to the rendering
b) fires fonts.loaded randomly
— I guess a) is a consequence of b —
we have no means to fix this programmatically. I made a guess that should:
a) fix navigation in Safari 10.1
b) does not reapply self.adapt() on every other browser
As it’s not that nice, it has to be reedited after Safari 11 is out and common. Hmm… when will this be?