BUG-700: Jump to paywall form from commentlink
We have no comments on a paywalled article, but a link to them. By adding an anchor to the paywall form, we now jump there, if we klick on the commentlink in the metadata. Maybe this will boost paywall registrations? Maybe we’ll need an a/b-test to find out.