ZON-4235: Align podcast link icons in mobile/desktop
Align podcast link icons in both mobile and desktop.
I had to remove justify-content in order to stack link icons right next
to the podlove button, while otherwise they would stick to the center.
Instead I used absolute positioning for the podlove link to the right
accordingly.
Please double check with stronger frontend fu if there is a more
pleasent solution to that.