BUG-812: Catch new zmo layout w/ another condition
As we try to put the camel through the needles eye once again aka. showing ads after elements the editorial staff can freely place anywhere by keeping the ads at the same position in the same time, we present: a new condition to watch what condition the region loop is in.