FIX: Adjust z.w.core.image.Image to LazyProxy's adapter behaviour
LazyProxy unrolls the adapter chain (e.g. uniqueId -> ICMSContent ->
IImages) only as far as possible, i.e. if one step fails, the
previously achieved result is returned as-is. I'm not really sure in
which cases this actually provides any benefits, but as its days are
hopefully numbered anyway due to TMS/ES, we're probably not going to
dig into that anymore.

Anyway, this means that IImages(content) can result in a
PersistenUnknownResource instance, which evaluates to true-ish, so we
need a more precise condition to handle this case.