SSL-221: Use a real request object in tests
The pyramid DummyRequest is too dissimilar to the real thing, (we've
noticed this before on several occasions). But now any tweaks we're
going to need for SSL capabilities would not work with it at all, so
it finally has to go. This is still kludgier than I'd like, but
honestly that's mostly on pyramid's testing setup, not ours.