ZON-4437: schema.org - Add new block type to wrap Q and A inside
FAQs consist of a question, represented by an intertitle block, and an
answer, represented by one or more blocks that are anything but an
intertitle (e.g. paragraphs, images, etc.).
schema.org optimization dictates, that each question and answer is
bundled together inside a seperate block. Therefore we need to loop
through all blocks, searching for intertitles and wrap them together with
all consecutive blocks until another intertitle is found or we have
reached the end of the page.