MAINT: Clarify reach API
The pattern "just write something onto an object" has the big drawback
that it's hard to reason about where the property came from, later on.
So let's rather be explicit about it, even though this adapter can
only ever be used in that _one_ specific code location/environment
(namely, on IReach result objects).