BUG-942: Retrieve (intextlink) body for ITMSContent
This is needed for teasers mostly so that the liveblog badge status
can be determined correctly.

With this change, ITMSContent results that want to access the body
incur an additional TMS request to retrieve it. The performance impact
of doing this for *every* teaser on a large CP is unclear, but luckily
that is not the case: only teasers for objects that provide
ILiveblogArticle currently want to look at the body, which is quite
the minority, so we should be alright.

On the other hand, the performance-related explanation in 8b2c373 that
disabled retrieving the TMS-body for ITMSContent is only half right:
While the body is indeed available from ES, it is currently not
included with the ES results (and the performance impact of doing
that, regarding the much larger data size per result object, is
unclear). Also, the body in ES is the raw CMS body without any
intextlinks filled-in, so it's not a drop-in replacement.