TMS-235: Support link objects in sitemaps
The production sitemaps include `blogpost` objects (and thus exclude
link objects, to avoid duplication). zeit.retresco treats indexed blog
objects as link objects (because that's the easiest way to present
them as ICMSContent), so our sitemap templates need to use the full
power of `create_url`, and not just rewrite the uniqueId (the
solr-wordpress plugin indexes the blogpost url as `uniqueId`, so our
previous code was... somewhat hacky).