ZON-4803: Simplify 18eedf6f and use it for sitemap tests.
The (mock) search results can now be set to either a string[*], in
which case it will be interpreted as a `uniqueId` and resolved into a
`ITMSRepresentation` (which is then used as the search result) or else
a dictionary containing the "full" mock search result, which will be
returned as is (except for the extra `url` key).

[*] Please note that for backward compatibility a dictionary with only a
`uniqueId` key will also be resolved.